Small Improvements in Processing Code
Some small improvements to the way images are processed...
As there is now tonemapping, the gamma values don't have to be so insanely high like before, so gamma is back to 2.2 as it should be (used to be at 3.15 to hide the ugly highlights). Doesn't look too much different, but improves final images I think.
Also more prepared to add the 3 way colour correction than it was before.

+++ platform/cocoa/main.m
@@ -224,9 +224,9 @@ int NSApplicationMain(int argc, const char * argv[])
     /* Link video with processing settings */
     setMlvProcessing(videoMLV, processingSettings);
     /* Limit frame cache to suitable amount of RAM (~33% at 8GB and below, ~50% at 16GB, then up and up) */
-    cacheSizeMB = (int)(0.66666f * (float)(MAC_RAM - 4000));
+    cacheSizeMB = (int)(0.66666 * (double)(MAC_RAM - 4000));
     if (MAC_RAM < 7500) cacheSizeMB = MAC_RAM * 0.33;
-    NSLog(@"Cache size = %iMB, or %i percent of RAM", cacheSizeMB, (int)((float)cacheSizeMB / (float)MAC_RAM * 100));
+    NSLog(@"Cache size = %iMB, or %i percent of RAM", cacheSizeMB, (int)((double)cacheSizeMB / (double)MAC_RAM * 100));
     setMlvRawCacheLimitMegaBytes(videoMLV, cacheSizeMB);
 
     /*
